Add hooks
The hooks style in our context is a list style.
Here the execution-ident part of the context finally makes sense.

  zstyle ":lookup:-browser-:*" hooks set_title

will run the function "set_title" at the start of the browser wrapper.
Might be useful for people, who want to set their screen or xterm titles
according to what lookup() does.